All leaks fixed

+++ Libft/src/ft_bzero.c
@@ -6,13 +6,13 @@
 /*   By: hcorrea- <hcorrea-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/15 15:28:32 by hcorrea-          #+#    #+#             */
-/*   Updated: 2023/01/13 10:24:00 by hcorrea-         ###   ########.fr       */
+/*   Updated: 2024/03/16 16:15:36 by hcorrea-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../inc/libft.h"
 
-void	ft_bzero(void *s, size_t n)
+void	ft_bzero(void *s, int n)
 {
 	ft_memset(s, '\0', n);
 }
