+++ Libft/src/ft_memchr.c
@@ -6,15 +6,15 @@
 /*   By: hcorrea- <hcorrea-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/14 17:07:05 by hcorrea-          #+#    #+#             */
-/*   Updated: 2023/01/13 10:24:43 by hcorrea-         ###   ########.fr       */
+/*   Updated: 2024/03/16 16:15:49 by hcorrea-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../inc/libft.h"
 
-void	*ft_memchr(const void *str, int c, size_t n)
+void	*ft_memchr(const void *str, int c, int n)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (i < n)
