+++ Libft/src/ft_memset.c
@@ -6,15 +6,15 @@
 /*   By: hcorrea- <hcorrea-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/14 16:27:43 by hcorrea-          #+#    #+#             */
-/*   Updated: 2023/01/13 10:25:02 by hcorrea-         ###   ########.fr       */
+/*   Updated: 2024/03/16 16:16:02 by hcorrea-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../inc/libft.h"
 
-void	*ft_memset(void *b, int c, size_t len)
+void	*ft_memset(void *b, int c, int len)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (i < len)
