+++ Libft/src/ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: hcorrea- <hcorrea-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/14 17:22:04 by hcorrea-          #+#    #+#             */
-/*   Updated: 2023/06/28 10:18:59 by hcorrea-         ###   ########.fr       */
+/*   Updated: 2024/03/16 16:10:16 by hcorrea-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../inc/libft.h"
 
-size_t	ft_strlen(const char *s)
+int	ft_strlen(const char *s)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (s[i])
